Small update toRBKI
